+++ plugins/woocommerce-admin/packages/components/src/text-control-with-affixes/index.js
@@ -3,9 +3,9 @@
  * External dependencies
  */
 import { Component } from '@wordpress/element';
+import { compose, withInstanceId } from '@wordpress/compose';
 import PropTypes from 'prop-types';
-import { BaseControl } from '@wordpress/components';
-import { withInstanceId } from '@wordpress/compose';
+import { BaseControl, withFocusOutside } from '@wordpress/components';
 import classnames from 'classnames';
 
 /**
@@ -14,6 +14,24 @@ import classnames from 'classnames';
  * a fixed part either at the beginning or at the end of the text input.
  */
 class TextControlWithAffixes extends Component {
+	constructor( props ) {
+		super( props );
+		this.state = {
+			isFocused: false,
+		};
+	}
+
+	handleFocusOutside() {
+		this.setState( { isFocused: false } );
+	}
+
+	handleOnClick( event, onClick ) {
+		this.setState( { isFocused: true } );
+		if ( typeof onClick === 'function' ) {
+			onClick( event );
+		}
+	}
+
 	render() {
 		const {
 			label,
@@ -22,12 +40,14 @@ class TextControlWithAffixes extends Component {
 			className,
 			instanceId,
 			onChange,
+			onClick,
 			prefix,
 			suffix,
 			type,
 			disabled,
 			...props
 		} = this.props;
+		const { isFocused } = this.state;
 
 		const id = `inspector-text-control-with-affixes-${ instanceId }`;
 		const onChangeValue = ( event ) => onChange( event.target.value );
@@ -42,6 +62,12 @@ class TextControlWithAffixes extends Component {
 			describedby.push( `${ id }__suffix` );
 		}
 
+		const baseControlClasses = classnames( className, {
+			'with-value': '' !== value,
+			empty: '' === value,
+			active: isFocused && ! disabled,
+		} );
+
 		const affixesClasses = classnames( 'text-control-with-affixes', {
 			'text-control-with-prefix': prefix,
 			'text-control-with-suffix': suffix,
@@ -49,7 +75,9 @@ class TextControlWithAffixes extends Component {
 		} );
 
 		return (
-			<BaseControl label={ label } id={ id } help={ help } className={ className }>
+			<BaseControl label={ label } id={ id } help={ help } className={ baseControlClasses }
+				onClick={ event => this.handleOnClick( event, onClick ) }
+			>
 				<div className={ affixesClasses }>
 					{ prefix && (
 						<span
@@ -68,6 +96,7 @@ class TextControlWithAffixes extends Component {
 						onChange={ onChangeValue }
 						aria-describedby={ describedby.join( ' ' ) }
 						disabled={ disabled }
+						onFocus={ () => this.setState( { isFocused: true } ) }
 						{ ...props }
 					/>
 
@@ -129,4 +158,7 @@ TextControlWithAffixes.propTypes = {
 	disabled: PropTypes.bool,
 };
 
-export default withInstanceId( TextControlWithAffixes );
+export default compose( [
+	withInstanceId,
+	withFocusOutside, // this MUST be the innermost HOC as it calls handleFocusOutside
+] )( TextControlWithAffixes );